requirements.txt
# Core Dependencies
langgraph>=0.0.24
groq>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LangChain (if you're using any LangChain components)
langchain-core>=0.1.0
langchain-groq>=0.1.0

# Web Framework (if you plan to build a web interface)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Development & Testing
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
pylint>=3.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Google API Auth
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0

# For diagram generation (if using diagram_generator.py)
graphviz>=0.20.1
pygraphviz>=1.12

# Enhanced Task Manager Dependencies
# Note: sqlite3 is built into Python, no need to install
# pathlib is built into Python 3.4+
# uuid is built into Python
# json is built into Python
# threading is built into Python
# collections is built into Python
# enum is built into Python (3.4+)
